# evmrs: remove feature stack-array and add feature unsafe-stack