--- a/rust/src/types/stack.rs
+++ b/rust/src/types/stack.rs
@@ -1,131 +1,16 @@
-#[cfg(feature = "stack-array")]
-use std::{cmp::min, mem::MaybeUninit};
+use std::cmp::min;
 
 use crate::types::{u256, FailStatus};
 
-#[cfg(feature = "stack-array")]
-#[derive(Debug)]
-pub struct Stack {
-    data: [MaybeUninit<u256>; 1024],
-    len: usize,
-}
-
-#[cfg(feature = "stack-array")]
-impl Stack {
-    pub fn new(inner: &[u256]) -> Self {
-        let len = min(inner.len(), 1024);
-        let mut s = Self {
-            data: [MaybeUninit::uninit(); 1024],
-            len,
-        };
-        // SAFETY:
-        // &[T] and &[MaybeUninit<T>] have the same layout.
-        // With nightly rust this could be replaced by MaybeUninit::copy_from_slice
-        s.data[..len].copy_from_slice(unsafe {
-            std::mem::transmute::<&[u256], &[std::mem::MaybeUninit<u256>]>(inner)
-        });
-        s
-    }
-
-    pub fn as_slice(&self) -> &[u256] {
-        // SAFETY:
-        // &[T] and &[MaybeUninit<T>] have the same layout and the first self.len elements are
-        // initialized.
-        unsafe {
-            std::mem::transmute::<&[std::mem::MaybeUninit<u256>], &[u256]>(&self.data[..self.len])
-        }
-    }
-
-    pub fn len(&self) -> usize {
-        self.len
-    }
-
-    pub fn push(&mut self, value: impl Into<u256>) -> Result<(), FailStatus> {
-        if self.len >= 1024 {
-            return Err(FailStatus::StackOverflow);
-        }
-
-        self.data[self.len] = MaybeUninit::new(value.into());
-        self.len += 1;
-        Ok(())
-    }
-
-    pub fn pop<const N: usize>(&mut self) -> Result<[u256; N], FailStatus> {
-        self.check_underflow(N)?;
-
-        self.len -= N;
-        let start = self.data.as_ptr() as *const u256;
-        // SAFETY:
-        // This does not wrap and the whole range from start to start + self.len is valid.
-        let pop_start = unsafe { start.add(self.len) };
-        // SAFETY:
-        // The the first self.len elements are initialized (invariant).
-        // `self.len` just got decremented by N, which means now that the first `self.len + N`
-        // elements are initialized. Therefore, it is safe to read N elements starting at index
-        // `self.len` as an array of length N and type u256.
-        Ok(unsafe { std::ptr::read(pop_start as *const [u256; N]) })
-    }
-
-    pub fn peek(&self) -> Option<&u256> {
-        if self.len == 0 {
-            None
-        } else {
-            let top = &self.data[self.len - 1];
-            // SAFETY:
-            // The first self.len elements are initialized.
-            let top = unsafe { std::mem::transmute::<&MaybeUninit<u256>, &u256>(top) };
-            Some(top)
-        }
-    }
-
-    pub fn swap_with_top(&mut self, nth: usize) -> Result<(), FailStatus> {
-        self.check_underflow(nth + 1)?;
-
-        let start = self.data.as_mut_ptr();
-        // SAFETY:
-        // This does not wrap and the whole range is valid.
-        let top = unsafe { start.add(self.len - 1) };
-        // SAFETY:
-        // This does not wrap and the whole range is valid.
-        let nth = unsafe { top.sub(nth) };
-        // SAFETY:
-        // top and nth are valid pointers into self.data.
-        unsafe {
-            std::ptr::swap(top, nth);
-        }
-        Ok(())
-    }
-
-    pub fn nth(&self, nth: usize) -> Result<u256, FailStatus> {
-        self.check_underflow(nth + 1)?;
-        let start = self.data.as_ptr() as *const u256;
-        // SAFETY:
-        // This does not wrap and the whole range is valid.
-        let nth = unsafe { start.add(self.len - 1 - nth) };
-        // SAFETY:
-        // nth is a valid pointer into self.data which points to one of the first self.len elements,
-        // which are all initialized, so the access is in bounds and it is safe so read the element
-        // as u256.
-        Ok(unsafe { *nth })
-    }
-
-    #[inline(always)]
-    fn check_underflow(&self, min_len: usize) -> Result<(), FailStatus> {
-        if self.len < min_len {
-            return Err(FailStatus::StackUnderflow);
-        }
-        Ok(())
-    }
-}
-
-#[cfg(not(feature = "stack-array"))]
 #[derive(Debug)]
 pub struct Stack(Vec<u256>);
 
-#[cfg(not(feature = "stack-array"))]
 impl Stack {
     pub fn new(inner: &[u256]) -> Self {
-        Self(Vec::from(inner))
+        let len = min(inner.len(), 1024);
+        let mut v = Vec::with_capacity(1024);
+        v.extend_from_slice(&inner[..len]);
+        Self(v)
     }
 
     pub fn as_slice(&self) -> &[u256] {
@@ -140,6 +25,12 @@ impl Stack {
         if self.0.len() >= 1024 {
             return Err(FailStatus::StackOverflow);
         }
+        #[cfg(feature = "unsafe-stack")]
+        // SAFETY:
+        // self.0 is initialized with capacity 1024 and never shrunk.
+        unsafe {
+            std::hint::assert_unchecked(self.0.capacity() == 1024);
+        }
         self.0.push(value.into());
         Ok(())
     }
@@ -147,8 +38,27 @@ impl Stack {
     pub fn swap_with_top(&mut self, nth: usize) -> Result<(), FailStatus> {
         self.check_underflow(nth + 1)?;
 
-        let len = self.0.len();
-        self.0.as_mut_slice().swap(len - 1, len - 1 - nth);
+        #[cfg(not(feature = "unsafe-stack"))]
+        {
+            let len = self.0.len();
+            self.0.swap(len - 1, len - 1 - nth);
+        }
+        #[cfg(feature = "unsafe-stack")]
+        {
+            let start = self.0.as_mut_ptr();
+            // SAFETY:
+            // This does not wrap and the whole range is valid.
+            let top = unsafe { start.add(self.len() - 1) };
+            // SAFETY:
+            // This does not wrap and the whole range is valid.
+            let nth = unsafe { top.sub(nth) };
+            // SAFETY:
+            // top and nth are valid pointers into the initialized part of the vector.
+            unsafe {
+                std::ptr::swap(top, nth);
+            }
+        }
+
         Ok(())
     }
 
@@ -168,7 +78,13 @@ impl Stack {
 
     pub fn nth(&self, nth: usize) -> Result<u256, FailStatus> {
         self.check_underflow(nth + 1)?;
-        Ok(self.0[self.0.len() - nth - 1])
+        #[cfg(not(feature = "unsafe-stack"))]
+        return Ok(self.0[self.0.len() - 1 - nth]);
+        #[cfg(feature = "unsafe-stack")]
+        // SAFETY:
+        // self.0.len() >= nth + 1 was checked in check_underflow.
+        // Therefore self.0.len() - 1 - nth is in bounds.
+        return Ok(*unsafe { self.0.get_unchecked(self.0.len() - 1 - nth) });
     }
 
     #[inline(always)]